Update README.md
fixed a couple typos and filled-in the text since I was here, then I edited the whole thing for my own clarity
fwiw in the third paragraph I'm not happy with: "Unfortunately this means one might accidentally install everything again and replicate their system.", since I think it's not clear what actions the user must take for this to occur and the word install doesn't give me an indication of what that means in Julia where you would add or load or use.  It could also indicate if this is primarily a memory problem and if so which memory, primary, secondary, something else? :cake: